Add files via upload
This strategy for playing black-jack is really basic. At each point in the game, there are only a finite number of possibilities ahead of us - based on whether we Hit or Stand. So we calculate the probabilities of these choices resulting in a Win and whichever is higher we go with that choice.